requirements.txt
grpcio
grpcio-tools
pika
nameko